[11.7.1] Comprehensive Update and Improvement of Addons for Pixel Plugin
Changes in this commit:
- Revised the plugin name and description to achieve greater consistency and clarity. This update ensures that the plugin's purpose and functionality are more accurately communicated to users.
- Implemented a new check for the Facebook Pixel plugin's active status. This addition enhances the reliability of the Addons for Pixel plugin by verifying the necessary prerequisite.
- Introduced an admin notice to alert users when the Facebook Pixel plugin is not active. This feature aids in troubleshooting and ensures that users are aware of the plugin's dependencies.
- Refined the function responsible for inserting the Facebook Pixel script. This improvement optimizes the script's integration and performance.
- Enhanced the functions tailored for different types of content, including pages, single posts, and WooCommerce products. These enhancements ensure that the plugin functions effectively across various content types.
- Added Open Graph meta information for single posts. This addition improves the plugin's social media integration, particularly enhancing content visibility and sharing capabilities on platforms like Facebook.

What is this change for?:
The purpose of this update is to elevate the Addons for Pixel plugin to version 1.7.1, incorporating a range of improvements and fixes. These changes are aimed at enhancing the plugin's usability, reliability, and compatibility with different types of content. By refining its functionality and expanding its features, we ensure that the plugin provides a more robust and user-friendly experience.

What security risks exist?:
No specific security risks were addressed in this commit. However, the enhancements and fixes contribute to the overall stability and reliability of the plugin, indirectly supporting a secure and efficient user experience.