Increment Addons for Pixel Plugin to New Version
Changes made in this commit:
- Modified the version number in the addons-for-pixel.php file, updating it to 1.7.2. This change reflects the latest version of the Addons for Pixel plugin, ensuring that the plugin's version number is accurate and up-to-date.

Purpose of This Change:
The update to version 1.7.2 is a routine incrementation of the Addons for Pixel plugin. This version change is essential for tracking the plugin's development progress, managing updates, and ensuring users have access to the latest features and improvements. Keeping the version number current is crucial for effective version control and user awareness of the plugin's evolution.

Security Considerations:
There are no security risks directly associated with this update. Changing the version number is a standard practice in software development and does not alter the functionality or security of the plugin itself. It is a straightforward administrative update that maintains the clarity and accuracy of the plugin's versioning system.